[arcane,accelerator] Déplace 'RunCommandImpl.h' dans 'internal'.
