+++ arcane/src/arcane/accelerator/core/srcs.cmake
@@ -32,6 +32,7 @@ set( ARCANE_SOURCES
   internal/IRunnerRuntime.h
   internal/AcceleratorCoreGlobalInternal.h
   internal/MemoryTracer.h
+  internal/RunCommandImpl.h
   internal/RunQueueImpl.h
   internal/ReduceMemoryImpl.h
 )
